refactor: extract artists router (PR-9, Phase 1)
